create toNumber as util function
