+++ src/expression/comparison/gt.ts
@@ -5,6 +5,7 @@
 
 import { Evaluable, Result } from '../../common/evaluable'
 import { isNumber } from '../../common/type-check'
+import { toNumber } from '../../common/util'
 import { Comparison } from '../comparison'
 
 // Operator key
@@ -31,8 +32,7 @@ export class GreaterThan extends Comparison {
    * {@link Comparison.comparison}
    */
   comparison (left: Result, right: Result): boolean {
-    const operand = (value: Result) =>
-      this.strict ? value : this.toNumber(value)
+    const operand = (value: Result) => (this.strict ? value : toNumber(value))
 
     if (isNumber(operand(left)) && isNumber(operand(right))) {
       return (left as number) > (right as number)
